+++ lib/src/bones_api_entity_db_memory.dart
@@ -921,6 +921,9 @@ class DBMemorySQLAdapter extends DBSQLAdapter<DBMemorySQLAdapterContext> {
 
     var relationshipTables = _findRelationshipTables(table);
 
+    fieldsTypes.removeWhere((key, value) =>
+        relationshipTables.any((r) => r.relationshipField == key));
+
     tableScheme = TableScheme(table,
         relationship: relationship != null,
         idFieldName: idFieldName,
@@ -1047,6 +1050,7 @@ class DBMemorySQLAdapter extends DBSQLAdapter<DBMemorySQLAdapterContext> {
           ref.targetField,
           ref.targetFieldType,
           relTargetField,
+          relationshipField: sourceField,
         );
       });
     }).toList();